refactor: created project actions package and moved the project actions interface and add component class into it
