+++ src/main/java/com/example/demo2/projectactions/ProjectActions.java
@@ -1,4 +1,4 @@
-package com.example.demo2;
+package com.example.demo2.projectactions;
 
 public interface ProjectActions {
     void performAction();
