+++ src/main/java/module-info.java
@@ -12,4 +12,6 @@ module com.example.demo2 {
     exports com.example.demo2.componentmodel;
     exports com.example.demo2.componentnode;
     opens com.example.demo2.componentmodel to javafx.fxml;
+    exports com.example.demo2.projectactions;
+    opens com.example.demo2.projectactions to javafx.fxml;
 }
